refactor(plugins):adjust the directory structure
