+++ packages/pc-plugins/src/grid/index.ts
@@ -1,6 +1,6 @@
 import { modifyCSS, createDom } from '@antv/dom-util';
 import Canvas from '@antv/g-base/lib/abstract/canvas';
-import { IGraph } from '../../interface/graph';
+import { IGraph } from '@antv/g6/es/interface/graph';
 import { ViewPortEventParam } from '@antv/g6-core/es/types';
 import Base from '../base';
