+++ packages/pc-plugins/src/minimap/index.ts
@@ -1,6 +1,6 @@
 import { Canvas as GCanvas } from '@antv/g-canvas';
 import { Canvas as GSVGCanvas } from '@antv/g-svg';
-import Base, { IPluginBaseConfig } from '../base';
+import Base, { IPluginBaseConfig } from '../../base';
 import { isString, isNil, each, debounce } from '@antv/util';
 import { createDom, modifyCSS } from '@antv/dom-util';
 import { Graph } from '@antv/g6';
