+++ packages/pc-plugins/src/timeBar/controllerBtn.ts
