+++ packages/pc-plugins/src/timeBar/handler.ts
