+++ packages/pc-plugins/src/timeBar/timeBarSlice.ts
@@ -14,7 +14,7 @@ import {
   PRE_STEP_BTN,
   TIMEBAR_CONFIG_CHANGE,
 } from './constant';
-import { IGraph } from '../../interface/graph';
+import { IGraph } from '@antv/g6/es/interface/graph';
 import { ShapeStyle } from '@antv/g6-core/es/types';
 
 const DEFAULT_SELECTEDTICK_STYLE = {
