+++ packages/pc-plugins/src/timeBar/trend.ts
