+++ packages/pc-plugins/src/timeBar/trendTimeBar.ts
@@ -3,7 +3,7 @@ import { get, size, assign, each } from '@antv/util';
 import Trend, { TrendCfg } from './trend';
 import Handler from './handler';
 import ControllerBtn, { ControllerCfg } from './controllerBtn';
-import { IGraph } from '../../interface/graph';
+import { IGraph } from '@antv/g6/es/interface/graph';
 import { ShapeStyle } from '@antv/g6-core/es/types';
 import {
   VALUE_CHANGE,
